# docs/swagger.yaml
definitions:
...
  response.ErrorCodeResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  response.TeamResponse:
    properties:
      description:
        type: string
      invitelink:
        type: string
      name:
        type: string
    type: object
...
  response.UserResponse:
    properties:
      name:
        type: string
      telegram_id:
        type: string
    type: object
  team.CreateTeamInput:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  team.InviteJoinRequest:
    properties:
      invite_code:
        type: string
      telegram_id:
        type: string
    required:
    - invite_code
    - telegram_id
    type: object
info:
  contact: {}
  title: Сервис для контроля задачами и встречами команды
paths:
...
  /team:
    delete:
      consumes:
      - application/json
      description: Удаляет команду и очищает связи со всеми участниками. Доступно
        только для владельца команды.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Команда успешно удалена
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 'Error: Только руководитель может удалить команду Code: ONLY_MANAGER_DELETE_TEAM,
            Error: Вы не являетесь владельцем команды Code: NOT_OWNER_OF_TEAM'
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
        "500":
          description: Ошибка при удалении команды
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Удаление команды
      tags:
      - team
    post:
      consumes:
      - application/json
      description: Создает команду, если запрос исходит от пользователя с ролью manager.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      - description: Данные команды
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/team.CreateTeamInput'
      produces:
      - application/json
      responses:
        "200":
          description: Информация о созданной команде
          schema:
            $ref: '#/definitions/response.TeamResponse'
        "400":
          description: Ошибка валидации или отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Доступ запрещен (не менеджер)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка создания команды
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создание команды
      tags:
      - team
    put:
      consumes:
      - application/json
      description: Обновляет название и описание команды. Доступно только для менеджеров.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      - description: Данные для обновления команды
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/team.CreateTeamInput'
      produces:
      - application/json
      responses:
        "200":
          description: Команда успешно обновлена
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка валидации или отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Доступ запрещен (не менеджер)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Error:Отсутствует команда у пользователя Code:USER_HAS_NO_TEAM,
            Error:Команда не найдена Code:TEAM_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
        "500":
          description: Ошибка при обновлении команды
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Изменение информации о команде
      tags:
      - team
  /team/invite:
    get:
      consumes:
      - application/json
      description: Возвращает ссылку для приглашения новых участников в команду. Доступно
        только для менеджеров.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL ссылки-приглашения
          schema:
            type: string
        "400":
          description: Отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Доступ запрещен (не менеджер)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Error:Отсутствует команда у пользователя Code:USER_HAS_NO_TEAM,
            Error:Команда не найдена Code:TEAM_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
      summary: Получение ссылки-приглашения
      tags:
      - team
  /team/join:
    post:
      consumes:
      - application/json
      description: Позволяет пользователю присоединиться к команде, используя пригласительный
        код.
      parameters:
      - description: Данные для присоединения к команде
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/team.InviteJoinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное присоединение к команде
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 'Error: Команда не найдена CODE: INVITE_CODE_INVALID, Error:
            Пользователь не найден. Зарегистрируйтесь через бота. CODE: USER_NOT_FOUND'
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
        "409":
          description: Вы уже присоединились к этой команде
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при присоединении к команде
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Присоединение к команде
      tags:
      - team
  /team/leave:
    get:
      consumes:
      - application/json
      description: Позволяет пользователю выйти из текущей команды
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Команда успешно покинута
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Manager не может просто так покинуть команду
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 'Error: Отсутствует команда у пользователя Code: USER_HAS_NO_TEAM'
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
        "500":
          description: Ошибка при попытке покинуть команду
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Покинуть команду
      tags:
      - team
  /team/members:
    get:
      consumes:
      - application/json
      description: Возвращает список всех участников команды, кроме текущего пользователя.
        Доступно только для менеджеров.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список участников команды
          schema:
            items:
              $ref: '#/definitions/response.UserResponse'
            type: array
        "400":
          description: Отсутствует telegram_id или ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Доступ запрещен (не менеджер)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Error:Отсутствует команда у пользователя Code:USER_HAS_NO_TEAM,
            Error:Команда не найдена Code:TEAM_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
        "500":
          description: Ошибка при получении участников команды
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение списка участников команды
      tags:
      - team
  /team/my:
    get:
      consumes:
      - application/json
      description: Возвращает данные о команде, к которой принадлежит пользователь
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о команде
          schema:
            $ref: '#/definitions/response.TeamResponse'
        "400":
          description: Отсутствует telegram_id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 'Error:Отсутствует команда у пользователя Сode:USER_HAS_NO_TEAM,
            Error: Команда не найдена Сode:TEAM_NOT_FOUND'
          schema:
            $ref: '#/definitions/response.ErrorCodeResponse'
      summary: Получение информации о своей команде
      tags:
      - team
...
swagger: "2.0"